+++ site/templates/Homepage/LiveExamples/styled.js
@@ -1,23 +1,43 @@
 import styled from 'styled-components';
-import { fontSize } from '~/styles/fonts';
-import { cols, ENTITY, SUPER_LIGHT_GREY, LIGHT_GREY, SKEW, UNSKEW } from '~/styles/vars';
+import { fontBold, fontSize } from '~/styles/fonts';
+import { cols, WHITE, ENTITY, SUPER_LIGHT_GREY, LIGHT_GREY, SKEW, UNSKEW } from '~/styles/vars';
+import { ExampleContainer } from '../USPs/styled';
 
 export const Container = styled.div`
   background: ${SUPER_LIGHT_GREY};
-  margin: ${cols(2)} 0;
+  margin: ${cols(2)} 0 ${cols(4)};
   display: flex;
-  overflow-x: hidden;
 `;
 
 export const LiveExampleContainer = styled.div`
   flex: 50%;
   display: flex;
-  align-items: center;
+  align-items: stretch;
   justify-content: flex-end;
 `;
 
 export const ExamplePositioning = styled.div`
   flex: 0 1 450px;
+  display: flex;
+`;
+
+export const AlignCenter = ExamplePositioning.extend`
+  align-items: center;
+  justify-content: center;
+`;
+
+export const VerticalCenter = ExamplePositioning.extend`
+  align-items: center;
+`;
+
+export const TopLeft = ExamplePositioning.extend`
+  justify-content: flex-start;
+  padding-top: ${cols(1)};
+`;
+
+export const BottomCenter = ExamplePositioning.extend`
+  align-items: flex-end;
+  justify-content: center;
 `;
 
 export const CodeContainer = styled.div`
@@ -37,4 +57,17 @@ export const Box = styled.div`
   width: 80px;
   height: 80px;
   background: ${ENTITY};
+`;
+
+export const Ball = styled.div`
+  width: 100px;
+  height: 100px;
+  background: ${ENTITY};
+  border-radius: 50%;
+  color: ${WHITE};
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  ${fontBold}
+  margin: 0 auto;
 `;